refactor(exercise_detail): split into per-widget files + layout value type
The biggest page in the codebase splits into:

detail_table_layout.dart  — DetailTableLayout value class (renamed
                            from the private _Layout) + EditField enum
                            shared between the page and its dialog.

widgets/:
- detail_app_bar.dart        DetailAppBar
- exercise_name_header.dart  ExerciseNameHeader
- table_header_row.dart      TableHeaderRow ("series / load / repeats")
- series_row.dart            SeriesRow (Slidable + Edit-load/repeats)
- series_row_body.dart       SeriesRowBody (the three cells, now also
                             reachable as a unit test target)
- body_parts_section.dart    BodyPartsSection (Primary / Secondary)
- body_part_blocks.dart      BodyPartBlocks
- set_value_dialog.dart      SetValueDialog (Edit load / Edit repeats)

SeriesRow now uses the shared DeleteActionPane for its start pane,
matching the work-log list. The page file shrinks from ~645 lines
to ~190 lines.
