+++ lib/presentation/pages/exercise_detail/exercise_detail_page.dart
@@ -1,10 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_slidable/flutter_slidable.dart';
 import 'package:workout_log/data/db/work_log_dao.dart';
-import 'package:workout_log/domain/models/body_part.dart';
 import 'package:workout_log/domain/models/work_log.dart';
-import 'package:workout_log/presentation/pages/exercise_form_page.dart';
+import 'package:workout_log/presentation/pages/exercise_detail/detail_table_layout.dart';
+import 'package:workout_log/presentation/pages/exercise_detail/widgets/body_parts_section.dart';
+import 'package:workout_log/presentation/pages/exercise_detail/widgets/detail_app_bar.dart';
+import 'package:workout_log/presentation/pages/exercise_detail/widgets/exercise_name_header.dart';
+import 'package:workout_log/presentation/pages/exercise_detail/widgets/series_row.dart';
+import 'package:workout_log/presentation/pages/exercise_detail/widgets/set_value_dialog.dart';
+import 'package:workout_log/presentation/pages/exercise_detail/widgets/table_header_row.dart';
+import 'package:workout_log/presentation/pages/exercise_form/exercise_form_page.dart';
 import 'package:workout_log/presentation/providers/data_providers.dart';
 import 'package:workout_log/presentation/theme/workout_colors.dart';
 import 'package:workout_log/presentation/util/responsive.dart';
@@ -25,7 +30,7 @@ class _ExerciseDetailPageState extends ConsumerState<ExerciseDetailPage> {
 
   // Local copy edited via copyWith; parent route refetches on pop.
   late WorkLog _workLog;
-  late _Layout _layout;
+  late DetailTableLayout _layout;
 
   WorkLogDao get _workLogDao => ref.read(workLogDaoProvider);
 
@@ -39,15 +44,15 @@ class _ExerciseDetailPageState extends ConsumerState<ExerciseDetailPage> {
   Widget build(BuildContext context) {
     return ResponsiveScaffold(
       appBarBuilder: (context, dims) {
-        _layout = _Layout.from(dims);
-        return _DetailAppBar(workLog: _workLog, layout: _layout);
+        _layout = DetailTableLayout.from(dims);
+        return DetailAppBar(workLog: _workLog, layout: _layout);
       },
       body: Builder(builder: (context) {
-        _layout = _Layout.from(ResponsiveDimensions.of(context));
+        _layout = DetailTableLayout.from(ResponsiveDimensions.of(context));
         final keys = _workLog.series.keys.toList();
         return Column(
           children: <Widget>[
-            _ExerciseNameHeader(
+            ExerciseNameHeader(
               workLog: _workLog,
               layout: _layout,
               onEdit: _openEditExercise,
@@ -55,10 +60,10 @@ class _ExerciseDetailPageState extends ConsumerState<ExerciseDetailPage> {
             Column(
               mainAxisAlignment: MainAxisAlignment.spaceAround,
               children: <Widget>[
-                _BodyPartsSection(workLog: _workLog, layout: _layout),
+                BodyPartsSection(workLog: _workLog, layout: _layout),
               ],
             ),
-            _TableHeaderRow(layout: _layout),
+            TableHeaderRow(layout: _layout),
             Divider(
               indent: _layout.screenWidth * 0.05,
               endIndent: _layout.screenWidth * 0.05,
@@ -84,15 +89,16 @@ class _ExerciseDetailPageState extends ConsumerState<ExerciseDetailPage> {
                       color: WorkoutColors.of(context).borderColor,
                     );
                   }
-                  return _SeriesRow(
+                  return SeriesRow(
                     index: keyIndex,
                     setKey: keys[keyIndex],
                     workLog: _workLog,
                     layout: _layout,
                     onDelete: () => _deleteSeries(keyIndex),
-                    onEditLoad: () => _openEditDialog(_EditField.load, keys[keyIndex]),
+                    onEditLoad: () =>
+                        _openEditDialog(EditField.load, keys[keyIndex]),
                     onEditRepeats: () =>
-                        _openEditDialog(_EditField.repeats, keys[keyIndex]),
+                        _openEditDialog(EditField.repeats, keys[keyIndex]),
                   );
                 },
               ),
@@ -120,13 +126,13 @@ class _ExerciseDetailPageState extends ConsumerState<ExerciseDetailPage> {
     );
   }
 
-  Future<void> _openEditDialog(_EditField field, String setKey) async {
+  Future<void> _openEditDialog(EditField field, String setKey) async {
     Util.blockOrientation(_layout.isPortrait);
     final result = await showDialog<String>(
       context: context,
-      builder: (_) => _SetValueDialog(
-        title: field == _EditField.load ? 'Edit load value' : 'Edit repeats number',
-        hint: field == _EditField.load
+      builder: (_) => SetValueDialog(
+        title: field == EditField.load ? 'Edit load value' : 'Edit repeats number',
+        hint: field == EditField.load
             ? _workLog.getLoad(setKey)
             : _workLog.getReps(setKey),
         isPortrait: _layout.isPortrait,
@@ -136,7 +142,7 @@ class _ExerciseDetailPageState extends ConsumerState<ExerciseDetailPage> {
     Util.unlockOrientation();
     if (!mounted || result == null) return;
 
-    final updated = field == _EditField.load
+    final updated = field == EditField.load
         ? _workLog.copyWith(load: {..._workLog.load, setKey: result})
         : _workLog.copyWith(series: {..._workLog.series, setKey: result});
     await _workLogDao.update(updated);
@@ -186,460 +192,4 @@ class _ExerciseDetailPageState extends ConsumerState<ExerciseDetailPage> {
   void _invalidateParent() {
     ref.invalidate(workLogsByDateProvider(_workLog.created));
   }
-}
-
-enum _EditField { load, repeats }
-
-class _DetailAppBar extends StatelessWidget implements PreferredSizeWidget {
-  const _DetailAppBar({required this.workLog, required this.layout});
-
-  final WorkLog workLog;
-  final _Layout layout;
-
-  @override
-  Size get preferredSize =>
-      Size.fromHeight(layout.isPortrait ? layout.screenHeight * 0.08 : layout.screenHeight * 0.1);
-
-  @override
-  Widget build(BuildContext context) {
-    final colors = WorkoutColors.of(context);
-    return AppBar(
-      title: Row(
-        mainAxisAlignment: MainAxisAlignment.end,
-        children: <Widget>[
-          SizedBox(
-            width: layout.screenWidth * 0.3,
-            child: Text(
-              workLog.created.toIso8601String().substring(0, 10),
-              textAlign: TextAlign.end,
-              style: TextStyle(color: colors.titleColor),
-            ),
-          ),
-        ],
-      ),
-      backgroundColor: colors.appBarColor,
-    );
-  }
-}
-
-class _ExerciseNameHeader extends StatelessWidget {
-  const _ExerciseNameHeader({
-    required this.workLog,
-    required this.layout,
-    required this.onEdit,
-  });
-
-  final WorkLog workLog;
-  final _Layout layout;
-  final VoidCallback onEdit;
-
-  @override
-  Widget build(BuildContext context) {
-    final colors = WorkoutColors.of(context);
-    return GestureDetector(
-      onLongPress: onEdit,
-      child: Container(
-        height: layout.isPortrait
-            ? layout.exerciseHeightPortrait
-            : layout.exerciseHeightLandscape,
-        width: layout.exerciseWidth,
-        alignment: const FractionalOffset(0.5, 0.5),
-        child: Text(
-          workLog.exercise.name,
-          textAlign: TextAlign.center,
-          style: TextStyle(
-            color: colors.textColor,
-            fontSize: WorkoutTypography.headerSize,
-            fontWeight: FontWeight.bold,
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _TableHeaderRow extends StatelessWidget {
-  const _TableHeaderRow({required this.layout});
-
-  final _Layout layout;
-
-  @override
-  Widget build(BuildContext context) {
-    final colors = WorkoutColors.of(context);
-    final height = layout.isPortrait
-        ? layout.portraitColumnHeight
-        : layout.headerLandscapeColumnHeight;
-    Widget cell(String text, double width) => Container(
-          height: height,
-          width: width,
-          alignment: const FractionalOffset(0.5, 0.5),
-          child: Text(
-            text,
-            style: TextStyle(
-              color: colors.textColor,
-              fontSize: WorkoutTypography.fontSize,
-              fontWeight: FontWeight.bold,
-            ),
-          ),
-        );
-    return Row(
-      children: <Widget>[
-        cell('series', layout.seriesColumnWidth),
-        cell('load', layout.columnWidth),
-        cell('repeats', layout.columnWidth),
-      ],
-    );
-  }
-}
-
-class _SeriesRow extends StatelessWidget {
-  const _SeriesRow({
-    required this.index,
-    required this.setKey,
-    required this.workLog,
-    required this.layout,
-    required this.onDelete,
-    required this.onEditLoad,
-    required this.onEditRepeats,
-  });
-
-  final int index;
-  final String setKey;
-  final WorkLog workLog;
-  final _Layout layout;
-  final VoidCallback onDelete;
-  final VoidCallback onEditLoad;
-  final VoidCallback onEditRepeats;
-
-  @override
-  Widget build(BuildContext context) {
-    final cellHeight = layout.isPortrait
-        ? layout.portraitColumnHeight
-        : layout.landscapeColumnHeight;
-    final actionMargin = EdgeInsets.symmetric(
-      vertical: layout.screenHeight * 0.01,
-      horizontal: layout.screenWidth * 0.01,
-    );
-
-    return Slidable(
-      key: ValueKey(setKey),
-      startActionPane: ActionPane(
-        motion: const ScrollMotion(),
-        extentRatio: 0.25,
-        children: [
-          Container(
-            margin: actionMargin,
-            child: SlidableAction(
-              onPressed: (_) => onDelete(),
-              backgroundColor: Colors.red,
-              foregroundColor: Colors.white,
-              icon: Icons.delete,
-              label: 'Delete',
-            ),
-          ),
-        ],
-      ),
-      endActionPane: ActionPane(
-        motion: const ScrollMotion(),
-        extentRatio: 0.25,
-        children: [
-          Container(
-            margin: actionMargin,
-            child: SlidableAction(
-              onPressed: (_) => onEditLoad(),
-              backgroundColor: Colors.yellow,
-              foregroundColor: Colors.black,
-              icon: Icons.edit,
-              label: 'Edit load',
-            ),
-          ),
-          Container(
-            margin: actionMargin,
-            child: SlidableAction(
-              onPressed: (_) => onEditRepeats(),
-              backgroundColor: Colors.green,
-              foregroundColor: Colors.white,
-              icon: Icons.edit,
-              label: 'Edit repeats',
-            ),
-          ),
-        ],
-      ),
-      child: _SeriesRowBody(
-        index: index,
-        load: workLog.getLoad(setKey),
-        reps: workLog.getReps(setKey),
-        cellHeight: cellHeight,
-        layout: layout,
-        onEditLoad: onEditLoad,
-        onEditRepeats: onEditRepeats,
-      ),
-    );
-  }
-}
-
-class _SeriesRowBody extends StatelessWidget {
-  const _SeriesRowBody({
-    required this.index,
-    required this.load,
-    required this.reps,
-    required this.cellHeight,
-    required this.layout,
-    required this.onEditLoad,
-    required this.onEditRepeats,
-  });
-
-  final int index;
-  final String load;
-  final String reps;
-  final double cellHeight;
-  final _Layout layout;
-  final VoidCallback onEditLoad;
-  final VoidCallback onEditRepeats;
-
-  @override
-  Widget build(BuildContext context) {
-    final colors = WorkoutColors.of(context);
-    final cellStyle = TextStyle(
-      color: colors.textColor,
-      fontSize: WorkoutTypography.fontSize,
-    );
-
-    Widget cell({required double width, required Widget child}) => Container(
-          height: cellHeight,
-          width: width,
-          alignment: const FractionalOffset(0.5, 0.5),
-          child: child,
-        );
-
-    return Row(
-      children: <Widget>[
-        cell(
-          width: layout.seriesColumnWidth,
-          // The column heading is "series", so the first row should
-          // read "1" — display the 1-based position rather than the
-          // zero-indexed array slot.
-          child: Center(child: Text((index + 1).toString(), style: cellStyle)),
-        ),
-        cell(
-          width: layout.columnWidth,
-          child: MaterialButton(
-            onPressed: onEditLoad,
-            child: Text(load, style: cellStyle),
-          ),
-        ),
-        cell(
-          width: layout.columnWidth,
-          child: MaterialButton(
-            onPressed: onEditRepeats,
-            child: Text(reps, style: cellStyle),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class _BodyPartsSection extends StatelessWidget {
-  const _BodyPartsSection({required this.workLog, required this.layout});
-
-  final WorkLog workLog;
-  final _Layout layout;
-
-  @override
-  Widget build(BuildContext context) {
-    final colors = WorkoutColors.of(context);
-    return Column(
-      children: <Widget>[
-        Text(
-          'Primary',
-          style: TextStyle(
-            color: colors.titleColor,
-            fontSize: layout.isPortrait
-                ? layout.titleFontSizePortrait
-                : layout.titleFontSizeLandscape,
-          ),
-        ),
-        SizedBox(height: layout.screenHeight * 0.01),
-        _BodyPartBlocks(parts: workLog.exercise.bodyParts, layout: layout),
-        SizedBox(height: layout.screenHeight * 0.02),
-        const Text('Secondary'),
-        SizedBox(height: layout.screenHeight * 0.01),
-        _BodyPartBlocks(
-            parts: workLog.exercise.secondaryBodyParts, layout: layout),
-      ],
-    );
-  }
-}
-
-class _BodyPartBlocks extends StatelessWidget {
-  const _BodyPartBlocks({required this.parts, required this.layout});
-
-  final Set<BodyPart> parts;
-  final _Layout layout;
-
-  @override
-  Widget build(BuildContext context) {
-    final colors = WorkoutColors.of(context);
-    final blocks = parts
-        .map((bp) => SizedBox(
-              height: layout.screenHeight * 0.05,
-              width: layout.screenWidth * 0.3,
-              child: Container(
-                color: Util.getBpColor(bp, colors),
-                child: Center(
-                  child: Text(
-                    Util.getBpName(bp),
-                    style: const TextStyle(color: Colors.amber),
-                  ),
-                ),
-              ),
-            ))
-        .toList();
-
-    if (blocks.length <= 3) {
-      return Row(
-        mainAxisAlignment: MainAxisAlignment.spaceAround,
-        children: blocks,
-      );
-    }
-    return Column(
-      children: <Widget>[
-        Row(
-            mainAxisAlignment: MainAxisAlignment.spaceAround,
-            children: blocks.take(3).toList()),
-        SizedBox(height: layout.screenHeight * 0.01),
-        Row(
-            mainAxisAlignment: MainAxisAlignment.spaceAround,
-            children: blocks.skip(3).toList()),
-      ],
-    );
-  }
-}
-
-class _SetValueDialog extends StatefulWidget {
-  const _SetValueDialog({
-    required this.title,
-    required this.hint,
-    required this.isPortrait,
-    required this.screenHeight,
-  });
-
-  final String title;
-  final String hint;
-  final bool isPortrait;
-  final double screenHeight;
-
-  @override
-  State<_SetValueDialog> createState() => _SetValueDialogState();
-}
-
-class _SetValueDialogState extends State<_SetValueDialog> {
-  final _controller = TextEditingController();
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  void _save() {
-    // int.parse throws on non-numeric input — that prevents saving an
-    // invalid value and surfaces a clear error.
-    final parsed = int.parse(_controller.text).toString();
-    Navigator.pop(context, parsed);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final colors = WorkoutColors.of(context);
-    final input = TextField(
-      controller: _controller,
-      autofocus: true,
-      autocorrect: true,
-      keyboardType: TextInputType.number,
-      decoration: InputDecoration(hintText: widget.hint),
-      maxLength: 4,
-    );
-    final actions = Row(
-      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-      children: <Widget>[
-        MaterialButton(
-          color: colors.greenButtonColor,
-          onPressed: _save,
-          child: Text('SAVE', style: TextStyle(color: colors.buttonTextColor)),
-        ),
-        MaterialButton(
-          color: colors.cancelButtonColor,
-          onPressed: () => Navigator.pop(context),
-          child:
-              Text('CANCEL', style: TextStyle(color: colors.buttonTextColor)),
-        ),
-      ],
-    );
-
-    if (widget.isPortrait) {
-      return SimpleDialog(
-        title: Center(heightFactor: 0.3, child: Text(widget.title)),
-        contentPadding: EdgeInsets.all(widget.screenHeight * 0.02),
-        children: [input, actions],
-      );
-    }
-    return SimpleDialog(
-      contentPadding: EdgeInsets.all(widget.screenHeight * 0.01),
-      children: <Widget>[
-        Center(heightFactor: 0.3, child: Text(widget.title)),
-        input,
-        actions,
-      ],
-    );
-  }
-}
-
-class _Layout {
-  const _Layout._({
-    required this.screenHeight,
-    required this.screenWidth,
-    required this.isPortrait,
-    required this.exerciseHeightPortrait,
-    required this.exerciseHeightLandscape,
-    required this.exerciseWidth,
-    required this.columnWidth,
-    required this.seriesColumnWidth,
-    required this.headerLandscapeColumnHeight,
-    required this.portraitColumnHeight,
-    required this.landscapeColumnHeight,
-    required this.titleFontSizePortrait,
-    required this.titleFontSizeLandscape,
-  });
-
-  factory _Layout.from(ResponsiveDimensions dims) => _Layout._(
-        screenHeight: dims.height,
-        screenWidth: dims.width,
-        isPortrait: dims.isPortrait,
-        exerciseHeightPortrait: dims.height * 0.1,
-        exerciseHeightLandscape: dims.height * 0.15,
-        exerciseWidth: dims.width,
-        columnWidth: dims.width * 0.375,
-        seriesColumnWidth: dims.width * 0.25,
-        headerLandscapeColumnHeight: dims.height * 0.15,
-        portraitColumnHeight: dims.height * 0.1,
-        landscapeColumnHeight: dims.height * 0.17,
-        titleFontSizePortrait: dims.width * 0.055,
-        titleFontSizeLandscape: dims.width * 0.03,
-      );
-
-  final double screenHeight;
-  final double screenWidth;
-  final bool isPortrait;
-  final double exerciseHeightPortrait;
-  final double exerciseHeightLandscape;
-  final double exerciseWidth;
-  final double columnWidth;
-  final double seriesColumnWidth;
-  final double headerLandscapeColumnHeight;
-  final double portraitColumnHeight;
-  final double landscapeColumnHeight;
-  final double titleFontSizePortrait;
-  final double titleFontSizeLandscape;
 }
